+++ WpfClient/Pages/Setting.xaml.cs
@@ -1,6 +1,4 @@
-﻿using ModernWpf.Controls;
-
-using System;
+﻿using System;
 using System.Collections.Generic;
 using System.Linq;
 using System.Text;
@@ -15,14 +13,14 @@ using System.Windows.Media.Imaging;
 using System.Windows.Navigation;
 using System.Windows.Shapes;
 
-namespace WpfClient
+namespace WpfClient.Pages
 {
     /// <summary>
-    /// Interaction logic for MainWindow.xaml
+    /// Setting.xaml の相互作用ロジック
     /// </summary>
-    public partial class MainWindow : Window
+    public partial class Setting : Page
     {
-        public MainWindow()
+        public Setting()
         {
             InitializeComponent();
         }
